backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
pydantic==1.10.13
python-multipart==0.0.6
httpx==0.25.2
loguru==0.7.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
gunicorn==21.2.0
motor==3.3.2
pymongo==4.6.3
